+++ UnitTests/EXPExpect+Test.m
@@ -1,0 +1,11 @@
+#import "EXPExpect+Test.h"
+#import "FakeTestCase.h"
+
+@implementation EXPExpect (Test)
+
+- (EXPExpect *)test {
+  self.testCase = [FakeTestCase new];
+  return self;
+}
+
+@end